#225: Don't pull Mongo binary on 'npm ci' on local
By default, mongodb-memory-server pulls in a MongoDB binary on install.
On a first run of 'npm ci', this makes npm seem to hang for no reason.
No progress bar or anything is shown either.

With this change, the MongoDB binary is only downloaded when needed.
If needed, the download happens during tests, with progress indication.
No download will happen for people using local Mongo for testing.